* Moved ACCelerator interface out of "include" just into "acc" (one level up), which signifies that these are the only common bits a backend needs to implement. Adjusted all relative include paths according to moved files. * Moved code shared/common between Cuda and HIP into acc/cuda directory (really it is using a Cuda API except for the prefix aka HIP). Made DBCSR_ACC_SRCS and artificial set pointing to either DBCSR_CUDA_SRCS, DBCSR_HIP_SRCS, or anything else. The previous location of the common code (common to Cuda/HIP) broke original idea of having separate backends that just implement acc.h. Adjusted all relative include paths according to moved files. Conditionally include HIP headers when referred outside of HIP-specific backend/folder. * Normalized header-inclusion style and grouped standard headers separately. The latter plus "local/own headers first" ensures our headers are self-contained and do not impose additional dependencies when potentially included outside of DBCSR project. Note clang-format is reordering header by default within a group (alphabetic order). * Changed ifdef to if-defined for consistency (combination of multiple conditions, etc).
