+++ src/acc/cuda/acc_error.cpp
@@ -6,32 +6,27 @@
  * For further information please visit https://dbcsr.cp2k.org                                    *
  * SPDX-License-Identifier: GPL-2.0+                                                              *
  *------------------------------------------------------------------------------------------------*/
-#ifndef DBCSR_ACC_LIBSMM_H
-#define DBCSR_ACC_LIBSMM_H
 
-#include "acc.h"
-
-#ifdef __cplusplus
-extern "C" {
+#if defined(__CUDA)
+# include "acc_cuda.h"
+#elif defined(__HIP)
+# include "../hip/acc_hip.h"
 #endif
 
-typedef struct libsmm_acc_stack_descriptor_type {
-  int m, n, k, max_m, max_n, max_k;
-  acc_bool_t defined_mnk;
-} libsmm_acc_stack_descriptor_type;
+#include "acc_error.h"
 
-int libsmm_acc_init(void);
-int libsmm_acc_is_thread_safe(void);
+#include <stdio.h>
+#include <math.h>
 
-int libsmm_acc_transpose(const int* dev_trs_stack, int offset, int nblks,
-  void* dev_data, acc_data_t datatype, int m, int n, acc_stream_t* stream);
-
-int libsmm_acc_process(const libsmm_acc_stack_descriptor_type* dev_param_stack, int stack_size,
-  int nparams, acc_data_t datatype, const void* dev_a_data, const void* dev_b_data, void* dev_c_data,
-  int m_max, int n_max, int k_max, acc_bool_t def_mnk, acc_stream_t* stream);
-
-#ifdef __cplusplus
+/****************************************************************************/
+int acc_error_check (ACC(Error_t) error){
+  if (error != ACC(Success)){
+      printf (BACKEND" error: %s\n", ACC(GetErrorString)(error));
+      return -1;
+    }
+  return 0;
 }
-#endif
 
-#endif /*DBCSR_ACC_LIBSMM_H*/
+extern "C" void acc_clear_errors () {
+  ACC(GetLastError)();
+}
