+++ src/acc/cuda/acc_error.h
@@ -7,10 +7,10 @@
  * SPDX-License-Identifier: GPL-2.0+                                                              *
  *------------------------------------------------------------------------------------------------*/
 
-#ifdef __CUDA
-#include "cuda/acc_cuda.h"
-#else
-#include "hip/acc_hip.h"
+#if defined(__CUDA)
+# include "acc_cuda.h"
+#elif defined(__HIP)
+# include "../hip/acc_hip.h"
 #endif
 
 int acc_error_check (ACC(Error_t) acc_error);
