+++ src/acc/cuda/acc_mem.cpp
@@ -7,16 +7,17 @@
  * SPDX-License-Identifier: GPL-2.0+                                                              *
  *------------------------------------------------------------------------------------------------*/
 
-#ifdef __CUDA
-#include "cuda/acc_cuda.h"
-#else
-#include "hip/acc_hip.h"
+#if defined(__CUDA)
+# include "acc_cuda.h"
+#elif defined(__HIP)
+# include "../hip/acc_hip.h"
 #endif
+
+#include "acc_error.h"
+#include "../acc.h"
 
 #include <stdio.h>
 #include <math.h>
-#include "acc_error.h"
-#include "include/acc.h"
 
 static const int verbose_print = 0;
