+++ src/acc/cuda/acc_stream.cpp
@@ -7,19 +7,20 @@
  * SPDX-License-Identifier: GPL-2.0+                                                              *
  *------------------------------------------------------------------------------------------------*/
 
-#ifdef __CUDA
-#include "cuda/acc_cuda.h"
-#else
-#include "hip/acc_hip.h"
+#if defined(__CUDA)
+# include "acc_cuda.h"
+#elif defined(__HIP)
+# include "../hip/acc_hip.h"
 #endif
+
+#include "acc_error.h"
+#include "../acc.h"
 
 #include <stdio.h>
 #include <math.h>
-#include "acc_error.h"
-#include "include/acc.h"
 
-#ifdef __CUDA_PROFILING
-#include <nvToolsExtCudaRt.h>
+#if defined(__CUDA_PROFILING)
+# include <nvToolsExtCudaRt.h>
 #endif
 
 static const int verbose_print = 0;
@@ -53,7 +54,7 @@ extern "C" int acc_stream_create(void** stream_p, const char* name, int priority
   if (acc_error_check(cErr)) return -1;
   if (acc_error_check(ACC(GetLastError)())) return -1;
 
-#ifdef __CUDA_PROFILING
+#if defined(__CUDA_PROFILING)
   nvtxNameCudaStreamA(*acc_stream, name);
 #endif
