--- src/acc/include/PACKAGE
@@ -1,5 +1,0 @@
-{
-"description": "Accelerator backend API and API for accelerated LIBSMM",
-"archive": "libdbcsr",
-"requires": []
-}
