+++ src/acc/libsmm_acc/kernels/smm_acc_common.h
@@ -7,10 +7,8 @@
  * SPDX-License-Identifier: GPL-2.0+                                                              *
  *------------------------------------------------------------------------------------------------*/
 
-#ifdef __HIP
-#if not defined(__HIP_PLATFORM_NVCC__)
-#include <hip/hip_runtime.h>
-#endif
+#if defined(__HIP) && !defined(__HIP_PLATFORM_NVCC__)
+# include <hip/hip_runtime.h>
 #endif
 
 #define MAX(x, y) (((x) > (y)) ? (x) : (y))
@@ -23,7 +21,7 @@
  * There is no native support for atomicAdd on doubles in Cuda 5.0. However   *
  * the following implementation is provided in the CUDA C Programing guide.   *
  ******************************************************************************/
-#ifdef __CUDA
+#if defined(__CUDA)
 #if (__CUDACC_VER_MAJOR__<8) || ( defined(__CUDA_ARCH__) && (__CUDA_ARCH__<600) )
 static __device__ double atomicAdd(double *address, double val) {
     unsigned long long int *address_as_ull =
@@ -43,7 +41,7 @@ static __device__ double atomicAdd(double *address, double val) {
 /******************************************************************************
  * A simple __ldg replacement for older cuda devices.                         *
  ******************************************************************************/
-#ifdef __CUDA
+#if defined(__CUDA)
 #if defined(__CUDA_ARCH__) && (__CUDA_ARCH__ < 350)
 #define __ldg(x)  (*(x))
 #endif
