+++ src/acc/libsmm_acc/libsmm_acc_benchmark.h
@@ -10,10 +10,10 @@
 #ifndef LIBSMM_ACC_BENCHMARK_H
 #define LIBSMM_ACC_BENCHMARK_H
 
-#ifdef __CUDA
-#include "../cuda/acc_cuda.h"
-#else
-#include "../hip/acc_hip.h"
+#if defined(__CUDA)
+# include "../cuda/acc_cuda.h"
+#elif defined(__HIP)
+# include "../hip/acc_hip.h"
 #endif
 
 #define MAX_BLOCK_DIM 80
